fix format string for mingw-w64

+++ Source/net_client.c
@@ -787,7 +787,7 @@ static void NET_CL_ParseGameData(net_packet_t *packet)
 
         recvobj->active = true;
         recvobj->cmd = cmd;
-        NET_Log("client: stored tic %lu in receive window", (unsigned long)seq + i);
+        NET_Log("client: stored tic %llu in receive window", (unsigned long long)seq + i);
 
         // If a packet is lost or arrives out of order, we might get
         // the tic in the next packet instead (because of extratic).
